split configuration into multiple tabs; support JSON editing (#74)
