+++ src/pages/tournament/[tournamentId]/config/plan.vue
