+++ src/pages/tournament/[tournamentId]/config/teams.vue
